Add files via upload
klib is a Python library used for importing, cleaning, analyzes, and preprocessing data. It is designed to simplify the data analysis process and provide a user-friendly interface for common data tasks, such as data profiling, data cleaning, data visualization, and data summarization. I have chosen the Telecom dataset for Exploratory data analysis and to explore the beautiful library for data analysis.